UPSTREAM: <carry>: remove kube import from crd openapi builder
Replace use of legacyscheme with hard-coded list of dotless built-in api
groups
